fix org setting visibility
